- Updated ImageMetricsService. - Fixed PaginatedResultDelegate related issues. - Updated ImagePlantResponse fields.

+++ images3-core/src/main/java/com/images3/core/ImagePlant.java
@@ -31,7 +31,7 @@ public interface ImagePlant {
     
     public Template fetchTemplate(String name);
     
-    public int countTemplates();
+    public long countTemplates();
     
     public PaginatedResult<List<Template>> listAllTemplates();
